refactor: repositories.py 5개 클래스 → 개별 파일 분리 (태스크 29.6)
- player_repository.py, room_repository.py, game_object_repository.py
- monster_repository.py, model_manager.py
- repositories.py 하위 호환성 re-export 유지
